# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk Sistem Reservasi Kampus (UniSpace).",
        "title": "Campus Reservation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Tim Developer",
            "email": "support@unu.ac.id"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/facilities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menampilkan daftar semua fasilitas yang tersedia.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facilities"
                ],
                "summary": "Lihat Semua Fasilitas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/facility.Facility"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan fasilitas baru lengkap dengan foto (Hanya Admin).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facilities"
                ],
                "summary": "Tambah Fasilitas Baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Fasilitas",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deskripsi",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lokasi",
                        "name": "location",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Kapasitas",
                        "name": "capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Harga per Jam",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Upload Foto (Max 4)",
                        "name": "photos",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/facilities/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melihat detail satu fasilitas berdasarkan ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facilities"
                ],
                "summary": "Detail Fasilitas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Fasilitas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/facility.Facility"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah data fasilitas (Full Update/PUT). Semua field teks WAJIB diisi ulang.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facilities"
                ],
                "summary": "Update Fasilitas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Fasilitas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nama Fasilitas (Wajib)",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deskripsi (Wajib)",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lokasi (Wajib)",
                        "name": "location",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Kapasitas (Wajib)",
                        "name": "capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Harga (Wajib)",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Foto Baru (Opsional, jika diisi akan menimpa yang lama)",
                        "name": "photos",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus fasilitas secara permanen (Hanya Admin).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facilities"
                ],
                "summary": "Hapus Fasilitas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Fasilitas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/facilities/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah status aktif fasilitas (Hanya Admin).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facilities"
                ],
                "summary": "Aktif/Nonaktifkan Fasilitas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Fasilitas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/facility.ToggleStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "facility.Facility": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "created_by_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "description": "UBAH: string -\u003e []string (Array)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "number"
                },
                "updated_by_name": {
                    "type": "string"
                }
            }
        },
        "facility.ToggleStatusReq": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format \"Bearer \u003ctoken_anda\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}